@@ -32,10 +32,10 @@
     "keycode": "^2.1.9",
     "query-string": "^5.0.1",
     "ramda": "^0.25.0",
-    "react": "^16.6.0",
+    "react": "^16.8.2",
     "react-addons-css-transition-group": "^15.6.2",
     "react-apollo": "^2.0.3",
-    "react-dom": "^16.6.0",
+    "react-dom": "^16.8.2",
     "react-feather": "^1.0.7",
     "react-redux": "^5.0.6",
     "react-router-dom": "^4.3.1",
@@ -58,8 +58,8 @@
     "@babel/polyfill": "^7.0.0",
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
-    "babel-plugin-styled-components": "^1.10.0",
     "babel-loader": "8.0.4",
+    "babel-plugin-styled-components": "^1.10.0",
     "clean-webpack-plugin": "^0.1.17",
     "dotenv": "^4.0.0",
     "file-loader": "^1.1.11",
